+++ merged-packages/bitcoin-wallet-snap/src/entities/currency.ts
@@ -8,6 +8,12 @@ export enum CurrencyUnit {
   Fiat = 'fiat', // Can also be cryptos like ETH, but will be fiat for 99% of users
 }
 
+export type CurrencyRate = {
+  conversionRate: number;
+  conversionDate: number;
+  currency: string;
+};
+
 export const networkToCurrencyUnit: Record<Network, CurrencyUnit> = {
   bitcoin: CurrencyUnit.Bitcoin,
   testnet: CurrencyUnit.Testnet,