+++ merged-packages/bitcoin-wallet-snap/src/infra/jsx/format.ts
@@ -1,8 +1,8 @@
 import type { Network } from '@metamask/bitcoindevkit';
 import { Amount, BdkErrorCode } from '@metamask/bitcoindevkit';
-import type { CaipAccountId, CurrencyRate } from '@metamask/snaps-sdk';
+import type { CaipAccountId } from '@metamask/snaps-sdk';
 
-import type { CurrencyUnit, Messages } from '../../entities';
+import type { CurrencyRate, CurrencyUnit, Messages } from '../../entities';
 import { networkToScope } from '../../handlers';
 
 export const displayAmount = (