+++ merged-packages/bitcoin-wallet-snap/CHANGELOG.md
@@ -15,6 +15,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Show a confirmation dialog before signing a PSBT from KeyringHandler and sending a transfer ([#591](https://github.com/MetaMask/snap-bitcoin-wallet/pull/591))
 - Add `resolveAccountAddress` method to KeyringHandler for dApp connectivity ([#590](https://github.com/MetaMask/snap-bitcoin-wallet/pull/590))
+- Bump `@metamask/keyring-api` from `^21.3.0` to `^22.0.0` ([#608](https://github.com/MetaMask/snap-bitcoin-wallet/pull/608))
+- Bump `@metamask/keyring-snap-sdk` from `^7.1.1` to `^8.0.0` ([#608](https://github.com/MetaMask/snap-bitcoin-wallet/pull/608))
+- Bump `@metamask/snaps-sdk` from `^10.3.0` to `^11.0.0` ([#608](https://github.com/MetaMask/snap-bitcoin-wallet/pull/608))
 
 ### Fixed